requirements.txt
Django==5.1.2
djangorestframework==3.15.2
psycopg2-binary==2.9.10
django-environ==0.11.2
celery==5.4.0
redis==5.0.8
django-redis==5.4.0
gunicorn==22.0.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1

